Move new classes to new builder package
