+++ poko-compiler-plugin/src/main/kotlin/dev/drewhamilton/poko/builder/GeneratedDeclarationsIrBodyFiller.kt
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package dev.drewhamilton.poko.ir
+package dev.drewhamilton.poko.builder
 
 import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
