+++ sample/jvm/src/test/kotlin/dev/drewhamilton/poko/sample/jvm/SampleBuilderTest.kt
@@ -1,8 +1,12 @@
 package dev.drewhamilton.poko.sample.jvm
 
+import assertk.assertFailure
 import assertk.assertThat
+import assertk.assertions.contains
 import assertk.assertions.isEqualTo
+import assertk.assertions.isNotNull
 import assertk.assertions.isNull
+import assertk.assertions.message
 import org.junit.Test
 
 class SampleBuilderTest {
@@ -36,15 +40,37 @@ class SampleBuilderTest {
     @Test fun `valid build succeeds`() {
         val sample: Sample = Sample.Builder()
             .setInt(1)
-            .setRequiredString("required string")
+            .setRequiredString("required")
             .build()
 
         assertThat(sample).isEqualTo(
             expected = Sample(
                 int = 1,
-                requiredString = "required string",
+                requiredString = "required",
                 optionalString = null,
             )
         )
+    }
+
+    @Test fun `invalid build fails`() {
+        // Missing int:
+        assertFailure {
+            Sample.Builder()
+                .setRequiredString("req")
+                .build()
+        }
+            .message()
+            .isNotNull()
+            .contains("is null")
+
+        // Missing string:
+        assertFailure {
+            Sample.Builder()
+                .setInt(66)
+                .build()
+        }
+            .message()
+            .isNotNull()
+            .contains("is null")
     }
 }
